==> 03_retroData/02_diffusion_scores.Rmd <==
---
title: "Diffusion scores"
author: "Sergio Picart-Armada"
date: "September 14, 2017"
output: html_document
---

```{r, message=FALSE}
library(igraph)
library(plyr)
library(magrittr)
library(diffuStats)

library(tictoc)

source("params.R")

# The kernel is heavy (1GB) and takes some time to load
load(file_kernel)

# the training data is mat.old
vars <- load(file_dataset)
stopifnot("mat.old" %in% vars)
```

# Compute the scores

```{r, message=FALSE}
set.seed(1)

# pagerank
pr <- igraph::page.rank(g.network)$vector

# use all genes as background (so that expected values are constant)

# try two versions of the network: mid-high and high confidence
list_kern <- list.files("~/big/devel/big/diffusion/", pattern = "biogrid.+rds") 
list_kern <- setNames(paste0("~/big/devel/big/diffusion/", list_kern), 
                      gsub("(biogrid_)(.*)(\\.rds)", "\\2", list_kern))

tic("Scoring algorithms")
plyr::l_ply(
  names(list_kern), 
  function(name_kern) {
    # kernel for a particular filtering setting
    K <- readRDS(list_kern[[name_kern]])
    
    # pick the labelled genes as a function of the background
    input <- mat.old*1
    # compute the scores
    list_diff <- lapply(
      list_methods, 
      function(method) {
        diffuStats::diffuse(
          K = K, 
          scores = input, 
          method = method, 
          n.perm = n.perm
        )
      }
    ) %>% set_names(list_methods)
    
    # Baseline: pagerank centralities
    list_diff$pagerank <- replicate(n = ncol(input), 
                                    expr = pr[rownames(K)]) %>%
      set_colnames(colnames(input))
        
    # dummy random scores
    set.seed(2)
    list_diff$random <- replicate(n = ncol(input), expr = sample(length(pr))) 
    dimnames(list_diff$random) <- dimnames(list_diff$pagerank) 
    
    # save with maximum compression
    save(list_diff, file = paste0(dir_scores, "/", name_kern, ".RData"), 
         compress = "xz")
  }
)
```

```{r}
# elapsed time
toc()
```


# Reproducibility

```{r}
out <- capture.output(sessionInfo())
writeLines(out, con = paste0(dir_metadata, "/02_sessionInfo.txt"))
```
